refactor(craft): extract plus-menu items builder into buildEntryMenuItems
Moves the pickerSections -> PlusMenuItem[] mapping out of CraftInputBar
into a pure, testable helper. CraftInputBar just calls it inside a memo.
